requirements.txt
# --- Core Framework & Services ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.4.0
redis==5.0.5
requests==2.32.3

# --- Database ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1

# --- Machine Learning & Data Processing ---
# Note: u8darts is a community fork of Darts with more frequent updates.
# Installing with [torch] ensures PyTorch is included.
u8darts[torch]==0.30.0
scikit-learn==1.5.0
lightgbm==4.3.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.0
# [Fix] Pin statsmodels and scipy to compatible versions to resolve ImportError
statsmodels==0.14.2
scipy==1.13.1
